LIBMAME.H is now public header file for MAME core
