+++ src/emu/uimenu.c
@@ -3916,7 +3916,7 @@ static void menu_select_game_build_driver_list(ui_menu *menu, select_game_state 
 	qsort((void *)menustate->driverlist, listnum, sizeof(menustate->driverlist[0]), menu_select_game_driver_compare_2);
     
     // pass driver list to myosd so it can create a custom GUI selection UI....
-    myosd_set_game_info((myosd_game_info**)menustate->driverlist, listnum);
+    myosd_set_game_info(menustate->driverlist, listnum);
 //DAV HACK
 }
