+++ src/osd/droid-ios/osdmain.c
@@ -30,7 +30,7 @@
 #include "osdinput.h"
 #include "osdsound.h"
 #include "osdvideo.h"
-#include "myosd-internal.h"
+#include "myosd.h"
 
 //============================================================
 //  GLOBALS
@@ -51,7 +51,6 @@ static const options_entry droid_mame_options[] =
 //  FUNCTION PROTOTYPES
 //============================================================
 
-//void osd_exit(running_machine *machine);
 static void osd_exit(running_machine &machine);
 
 //============================================================
@@ -68,50 +67,11 @@ int iOS_main  (int argc, char **argv)
 int main(int argc, char **argv)
 #endif
 {
-	static char *args[255];	int n=0;
-	int ret;
-	FILE *f;
-    
-	//printf("Iniciando\n");
-    
 	myosd_init();
 
-    // the iOS app will "re-run" MAME by calling into main again, dont loop here
-	//while(1)
-	{
-		droid_ios_setup_video();
-        
-        // cli_execute does the heavy lifting; if we have osd-specific options, we
-        // would pass them as the third parameter here
-		n=0;
-        if (argc == 0) {
-            args[n]= (char *)"mame4x";n++;
-        }
-        else {
-            while (n < argc) {
-                args[n]= argv[n];n++;
-            }
-        }
+    droid_ios_setup_video();
 
-		//args[n]= (char *)"starforc"; n++;
-		//args[n]= (char *)"1944"; n++;
-		//args[n]= (char *)"mslug3"; n++;
-        //args[n]= (char *)"dino"; n++;
-		//args[n]= (char *)"outrun"; n++;
-		//args[n]= (char *)"-autoframeskip"; n++;
-		//args[n]= (char *)"-noautoframeskip"; n++;
-		//args[n]= (char *)"-nosound"; n++;
-		//args[n]= (char *)"-novideo"; n++;
-		//args[n]= (char *)"-nosleep"; n++;
-        //args[n]= (char *)"-autosave"; n++;
-		//args[n]= (char *)"-sleep"; n++;
-		//args[n]= (char *)"-jdz"; n++;args[n]= (char *)"0.0"; n++;
-		//args[n]= (char *)"-jsat"; n++;args[n]= (char *)"1.0"; n++;
-		//args[n]= (char *)"-joystick_deadzone"; n++;args[n]= (char *)"0.0"; n++;
-		args[n]= (char *)"-nocoinlock"; n++;
-        
-        ret = cli_execute(n, args, droid_mame_options);
-	}
+    int ret = cli_execute(argc, argv, droid_mame_options);
     
 	myosd_deinit();
